requirements.txt
grpcio==1.59.0
grpcio-tools==1.59.0
pika==1.3.2
protobuf==4.24.4
psutil==5.9.6
numpy==1.26.1
